+++ configure.py
@@ -1001,6 +1001,7 @@ scylla_core = (['message/messaging_service.cc',
                 'db/virtual_tables.cc',
                 'db/system_distributed_keyspace.cc',
                 'db/size_estimates_virtual_reader.cc',
+                'db/schema_applier.cc',
                 'db/schema_tables.cc',
                 'db/cql_type_parser.cc',
                 'db/legacy_schema_migrator.cc',
