+++ db/CMakeLists.txt
@@ -7,6 +7,7 @@ target_sources(db
     virtual_tables.cc
     system_distributed_keyspace.cc
     size_estimates_virtual_reader.cc
+    schema_applier.cc
     schema_tables.cc
     cql_type_parser.cc
     legacy_schema_migrator.cc
