+++ db/schema_applier.hh
@@ -1,0 +1,27 @@
+/*
+ * Modified by ScyllaDB
+ * Copyright (C) 2024-present ScyllaDB
+ */
+
+/*
+ * SPDX-License-Identifier: (AGPL-3.0-or-later and Apache-2.0)
+ */
+
+#pragma once
+
+#include "mutation/mutation.hh"
+#include "seastar/core/future.hh"
+#include "service/storage_proxy.hh"
+#include "query-result-set.hh"
+
+#include <seastar/core/distributed.hh>
+
+namespace db {
+
+namespace schema_tables {
+
+future<> merge_schema(sharded<db::system_keyspace>& sys_ks, distributed<service::storage_proxy>& proxy, gms::feature_service& feat, std::vector<mutation> mutations, bool reload = false);
+
+}
+
+}
